Support to show mouse cursor

+++ src/kernel/setjmp.h
@@ -11,8 +11,8 @@ typedef long long jmp_buf[16];
 #define setjmp _setjmp
 #define longjmp _longjmp
 
-intptr_t _setjmp(jmp_buf);
-void _longjmp(jmp_buf, intptr_t) __attribute__((noreturn));
+int _setjmp(jmp_buf);
+void _longjmp(jmp_buf, int) __attribute__((noreturn));
 
 #ifdef __cplusplus
 }
